Use custom emojione types
