--- apps/meteor/app/emoji-emojione/lib/emojioneRender.ts
@@ -1,9 +1,0 @@
-import emojione from 'emojione';
-
-export function emojioneRender(message: string): string {
-	return emojione.toImage(message);
-}
-
-export function emojioneRenderFromShort(message: string): string {
-	return emojione.shortnameToImage(message);
-}
