+++ apps/meteor/package.json
@@ -104,7 +104,6 @@
 		"@types/cssom": "^0.4.1",
 		"@types/dompurify": "^2.3.3",
 		"@types/ejson": "^2.2.0",
-		"@types/emojione": "^2.2.6",
 		"@types/express": "^4.17.17",
 		"@types/express-rate-limit": "^5.1.3",
 		"@types/fibers": "^3.1.1",
